Poc for magic connect
