+++ packages/auth-kit/example/package.json
@@ -15,7 +15,8 @@
     "@safe-global/safe-react-components": "^2.0.1",
     "ethers": "^6.0.3",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0"
   },
   "devDependencies": {
     "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
