+++ packages/auth-kit/example/src/AppBar.tsx
@@ -1,38 +1,20 @@
-import { AppBar as MuiAppBar, Typography, styled, Box, Button } from '@mui/material'
-import { SafeGetUserInfoResponse, Web3AuthModalPack } from '../../src'
+import { Link as RouterLink } from 'react-router-dom'
+import { AppBar as MuiAppBar, Typography, styled, Link } from '@mui/material'
 
-type AppBarProps = {
-  isLoggedIn: boolean
-  onLogin: () => void
-  onLogout: () => void
-  userInfo?: SafeGetUserInfoResponse<Web3AuthModalPack>
-}
-
-const AppBar = ({ isLoggedIn, onLogin, onLogout, userInfo }: AppBarProps) => {
+const AppBar = () => {
   return (
     <StyledAppBar position="static" color="default">
-      <Typography variant="h3" pl={4} fontWeight={700}>
-        Auth Provider Demo
+      <Typography variant="h3" pl={3} fontWeight={700}>
+        Auth Kit
       </Typography>
-
-      <Box mr={5}>
-        {isLoggedIn ? (
-          <Box display="flex" alignItems="center">
-            {userInfo && (
-              <Typography variant="body1" fontWeight={700}>
-                Hello {userInfo.name || userInfo.email} !!
-              </Typography>
-            )}
-            <Button variant="contained" onClick={onLogout} sx={{ ml: 2 }}>
-              Log Out
-            </Button>
-          </Box>
-        ) : (
-          <Button variant="contained" onClick={onLogin}>
-            Login
-          </Button>
-        )}
-      </Box>
+      <nav>
+        <Link to={`/web3auth`} component={RouterLink} pl={2} sx={{ textDecoration: 'none' }}>
+          Web3Auth Modal
+        </Link>
+        <Link to={`/magic-connect`} component={RouterLink} pl={2} sx={{ textDecoration: 'none' }}>
+          Magic Connect
+        </Link>
+      </nav>
     </StyledAppBar>
   )
 }
@@ -44,7 +26,6 @@ const StyledAppBar = styled(MuiAppBar)`
     background: ${({ theme }) => theme.palette.background.paper};
     height: 70px;
     align-items: center;
-    justify-content: space-between;
     flex-direction: row;
     border-bottom: 2px solid ${({ theme }) => theme.palette.background.paper};
     box-shadow: none;
