+++ packages/auth-kit/src/index.ts
@@ -1,5 +1,8 @@
 export * from './packs/web3auth/Web3AuthModalPack'
 export * from './packs/web3auth/types'
 
+export * from './packs/magic/MagicConnectPack'
+export * from './packs/magic/types'
+
 export * from './AuthKitBasePack'
 export * from './types'
